모듈화 후, 추천기능 버그 생겨서 수정함
서비스 로직 누락되서 수정함.

- 게시글 삭제시 db오류 나는데, 확인하고 수정하기

- 메인화면 명령어 기능 구현하기

+++ src/main/java/com/sc1hub/board/AbstractBoardService.java
@@ -3,10 +3,11 @@ package com.sc1hub.board;
 import com.sc1hub.mapper.BoardMapper;
 import com.sc1hub.util.PageDTO;
 import com.sc1hub.util.PageService;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-public abstract class AbstractBoardService<T, C, R> implements BoardService<T, C, R> {
+public abstract class AbstractBoardService<T, C, R extends RecommendDTO> implements BoardService<T, C, R> {
 
     protected abstract BoardMapper<T, C, R> getBoardMapper();
 
@@ -109,14 +110,41 @@ public abstract class AbstractBoardService<T, C, R> implements BoardService<T, C
     }
 
     @Override
+    @Transactional
     public void addRecommendation(R recommendDTO) {
-        getBoardMapper().insertRecommendation(recommendDTO);
+        // 1. 사용자가 이미 해당 게시글을 추천했는지 확인
+        int count = getBoardMapper().checkRecommendation(recommendDTO);
+        if (count == 0) {
+            // 2. 추천하지 않았다면, 추천 테이블에 데이터를 추가하고, 게시글의 추천 수를 증가
+            getBoardMapper().insertRecommendation(recommendDTO);
+            // 3. 게시글의 총 추천 수를 갱신
+            updateRecommendCount(recommendDTO.getPostNum());
+        }
     }
 
     @Override
-    public void cancelRecommendation(R recommendDTO) {
+    @Transactional
+    public void cancelRecommendation(R recommendDTO) throws Exception {
+        // 1. 사용자가 이미 추천을 했는지 확인
+        int recommendCount = getBoardMapper().checkRecommendation(recommendDTO);
+        if (recommendCount == 0) {
+            throw new RuntimeException("해당 게시글에 대한 추천이 없습니다.");
+        }
+        // 2. 추천을 했다면, 해당 추천을 데이터베이스에서 삭제
         getBoardMapper().deleteRecommendation(recommendDTO);
+        // 3. 게시글의 총 추천 수를 갱신
+        updateRecommendCount(recommendDTO.getPostNum());
     }
+
+    private void updateRecommendCount(int postNum) {
+        int actualRecommendCount = getBoardMapper().getActualRecommendCount(postNum);
+        int currentRecommendCount = getBoardMapper().getRecommendCount(postNum);
+
+        if (actualRecommendCount != currentRecommendCount) {
+            getBoardMapper().updateTotalRecommendCount(postNum);
+        }
+    }
+
 
     @Override
     public int checkRecommendation(R recommendDTO) {
@@ -133,7 +161,6 @@ public abstract class AbstractBoardService<T, C, R> implements BoardService<T, C
         checkPageAndKeyword(page);
         return utilLoadingForCommentPage(page);
     }
-
 
 }
